Bugfix: XFrameOptionsFilter was renamed to SecurityHeadersFilter
(and generalised a bit, see the new CSP functionality)

This was a breaking change, and hard to track down!  It didn't show up
in the logs at all, even with DEVELOPER_LOGGING (everything set to
trace) turned on... perhaps that only shows up in Tomcat logging,
which in hindsight I didn't try.  I discovered this after running it
locally under Jetty.

Ref: https://github.com/geoserver/geoserver/pull/7514/files